+++ src/main/java/roomescape/controller/ReservationController.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.requestDto.ReservationCreateRequest;
-import roomescape.responseDto.AvailableTimeResponse;
-import roomescape.responseDto.ReservationResponse;
+import roomescape.dto.request.ReservationCreateRequest;
+import roomescape.dto.response.AvailableTimeResponse;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 @RestController
