+++ src/main/java/roomescape/dto/request/ReservationCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.requestDto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
