+++ src/main/java/roomescape/dto/request/ThemeCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.requestDto;
+package roomescape.dto.request;
 
 public record ThemeCreateRequest(String name, String description, String thumbnail) {
 }
