+++ src/main/java/roomescape/dto/response/AvailableTimeResponse.java
@@ -1,8 +1,7 @@
-package roomescape.responseDto;
+package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 
 public record AvailableTimeResponse(
         Long timeId,
