+++ src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -1,8 +1,7 @@
-package roomescape.responseDto;
+package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import roomescape.domain.Reservation;
 
 public record ReservationResponse(
