+++ src/main/java/roomescape/dto/response/ReservationTimeResponse.java
@@ -1,8 +1,7 @@
-package roomescape.responseDto;
+package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import roomescape.domain.ReservationTime;
 
 public record ReservationTimeResponse(
