+++ src/main/java/roomescape/service/ReservationService.java
@@ -13,9 +13,9 @@ import roomescape.domain.UserName;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.requestDto.ReservationCreateRequest;
-import roomescape.responseDto.AvailableTimeResponse;
-import roomescape.responseDto.ReservationResponse;
+import roomescape.dto.request.ReservationCreateRequest;
+import roomescape.dto.response.AvailableTimeResponse;
+import roomescape.dto.response.ReservationResponse;
 
 @Service
 public class ReservationService {
