+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationTimeRepository;
-import roomescape.requestDto.ReservationTimeCreateRequest;
-import roomescape.responseDto.ReservationTimeResponse;
+import roomescape.dto.request.ReservationTimeCreateRequest;
+import roomescape.dto.response.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
